+++ src/extensions/rules/rulesext.h
@@ -275,4 +275,15 @@ public:
 	 *  Determines whether Free Radar would still apply even during Low Power
 	 */
 	bool IsFreeRadarOnLowPower;
+
+    /**
+     *  Determines whether bridges should use the Bridge Health Tracking feature. 
+     *  When disabled, uses the vanilla bridge destruction logic (by random chance).
+     */
+    bool IsUseBridgeHealth;
+
+	/*
+    * The armor type used by bridges for damage calculation. Only used when Bridge Health mechanism is enabled.
+    */
+    ArmorType BridgeArmor;
 };